+++ .gitignore
@@ -1,0 +1,10 @@
+.venv/
+.pytest_cache/
+__pycache__/
+*.py[cod]
+.coverage
+htmlcov/
+*.lock
+
+.idea/
+.vscode/
