+++ pyinject/__init__.py
@@ -1,0 +1,4 @@
+from .decorators import AutoWired
+from .functions import Depends, run_program
+
+__all__ = ["AutoWired", "Depends", "run_program"]
